UI restyle

+++ UI/WebServerResources/MailerUI.js
@@ -379,7 +379,7 @@ function onDocumentKeydown(event) {
             }
         }
 	if (keyCode == Event.KEY_DELETE ||
-            keyCode == Event.KEY_BACKSPACE) {
+            keyCode == Event.KEY_BACKSPACE && isMac()) {
             deleteSelectedMessages();
             Event.stop(event);
         }
@@ -637,7 +637,7 @@ function onMailboxTreeItemClick(event) {
     }
     else {
         var datatype = this.parentNode.getAttribute("datatype");
-        if (datatype == 'draft' || datatype == 'sent')
+        if (datatype == 'draft' || datatype == 'draft/folder' || datatype == 'sent' || datatype == 'sent/folder')
             toggleAddressColumn("from", "to");
         else
             toggleAddressColumn("to", "from");
@@ -1249,22 +1249,36 @@ function loadMessage(msguid) {
     return seenStateHasChanged;
 }
 
-/**
- * Hide the "Load Images" button when there's no unsafe content
-*/
 function configureLoadImagesButton() {
+    // We show/hide the "Load Images" button
     var loadImagesButton = $("loadImagesButton");
+    var content = $("messageContent");
+    var hiddenImgs = [];
+    var imgs = content.select("IMG");
+    $(imgs).each(function(img) {
+            var unsafeSrc = img.getAttribute("unsafe-src");
+            if (unsafeSrc) {
+                hiddenImgs.push(img);
+            }
+        });
+    content.hiddenImgs = hiddenImgs;
+
+    var hiddenObjects = [];
+    var objects = content.select("OBJECT");
+    $(objects).each(function(obj) {
+            if (obj.getAttribute("unsafe-data")
+                || obj.getAttribute("unsafe-classid")) {
+                hiddenObjects.push(obj);
+            }
+        });
+    content.hiddenObjects = hiddenObjects;
+
     if (typeof(loadImagesButton) == "undefined" ||
         loadImagesButton == null ) {
         return;
     }
-    var content = $("messageContent");
-    var unsafeElements = content.select('[unsafe-src], [unsafe-data], [unsafe-classid], [unsafe-background]');
-    if (unsafeElements.length == 0) {
+    if ((hiddenImgs.length + hiddenObjects.length) == 0) {
         loadImagesButton.setStyle({ display: 'none' });
-    }
-    else {
-        content.hiddenElements = unsafeElements;
     }
 }
 
@@ -1343,27 +1357,13 @@ function configureLinksInMessage() {
             anchor.observe("contextmenu", onEmailAddressClick);
             anchor.writeAttribute("moz-do-not-send", false);
         }
-        else if (!anchor.id)
+        else
             anchor.observe("click", onMessageAnchorClick);
     }
 
-    var attachmentsMenu = $("attachmentsMenu");
-    if (attachmentsMenu) {
-        var options = attachmentsMenu.select("li");
-        var callbacks = [];
-        for (var i = 0; i < options.length; i++) {
-            if (options[i].className == 'separator')
-                callbacks.push(null);
-            else
-                callbacks.push(saveAttachment);
-        }
-        initMenu(attachmentsMenu, callbacks);
-        $("attachmentsHref").on("click", function (event) {
-            popupMenu(event, 'attachmentsMenu', this);
-            preventDefault(event);
-            return false;
-        });
-    }
+    var attachments = messageDiv.select ("DIV.linked_attachment_body");
+    for (var i = 0; i < attachments.length; i++)
+        $(attachments[i]).observe("contextmenu", onAttachmentClick);
 
     var images = messageDiv.select("IMG.mailer_imagecontent");
     for (var i = 0; i < images.length; i++)
@@ -1371,11 +1371,12 @@ function configureLinksInMessage() {
 
     var editDraftButton = $("editDraftButton");
     if (editDraftButton)
-        editDraftButton.on("click", onMessageEditDraft);
+        editDraftButton.observe("click",
+                                onMessageEditDraft.bindAsEventListener(editDraftButton));
 
     var loadImagesButton = $("loadImagesButton");
     if (loadImagesButton)
-        loadImagesButton.on("click", onMessageLoadImages);
+        $(loadImagesButton).observe("click", onMessageLoadImages);
 
     configureiCalLinksInMessage();
 }
@@ -1561,29 +1562,34 @@ function onMessageContentMenu(event) {
 }
 
 function onMessageEditDraft(event) {
-    Event.stop(event);
     return openMessageWindowsForSelection("edit", true);
 }
 
 function onMessageLoadImages(event) {
-    Event.stop(event);
     loadRemoteImages();
+    Event.stop(event);
 }
 
 function loadRemoteImages() {
     var content = $("messageContent");
-    if (content.hiddenElements) {
-        $(content.hiddenElements).each(function(element) {
-            ['src', 'data', 'classid', 'background'].each(function(attr) {
-                var unsafeAttr = element.readAttribute('unsafe-' + attr);
-                if (unsafeAttr) {
-                    log ('unsafe ' +  attr + ': ' + unsafeAttr);
-                    element.writeAttribute(attr, unsafeAttr);
-                }
-            });
+    $(content.hiddenImgs).each(function(img) {
+            var unSafeSrc = img.getAttribute("unsafe-src");
+            log ("unsafesrc: " + unSafeSrc);
+            img.src = img.getAttribute("unsafe-src");
         });
-        content.hiddenElements = null;
-    }
+    content.hiddenImgs = null;
+    $(content.hiddenObjects).each(function(obj) {
+            var unSafeData = obj.getAttribute("unsafe-data");
+            if (unSafeData) {
+                obj.setAttribute("data", unSafeData);
+            }
+            var unSafeClassId = obj.getAttribute("unsafe-classid");
+            if (unSafeClassId) {
+                obj.setAttribute("classid", unSafeClassId);
+            }
+        });
+    content.hiddenObjects = null;
+
     var loadImagesButton = $("loadImagesButton");
     if (loadImagesButton)
         loadImagesButton.setStyle({ display: 'none' });
@@ -1604,6 +1610,12 @@ function onMessageAnchorClick(event) {
 function onImageClick(event) {
     popupMenu(event, 'imageMenu', this);
     preventDefault(event);
+    return false;
+}
+
+function onAttachmentClick (event) {
+    popupMenu (event, 'attachmentMenu', this);
+    preventDefault (event);
     return false;
 }
 
@@ -1761,29 +1773,13 @@ function saveImage(event) {
     window.location.href = urlAsAttachment;
 }
 
-/* Download a file using a temporary iframe that we delete once the download is started */
-function download(url) {
-    var form = createElement('form', null, 'hidden', { action: url, method: 'GET'});
-    $(document.body).appendChild(form);
-    var div = AIM.submit(form);
-    form.submit();
-    setTimeout(function () {
-        form.remove();
-        div.remove();
-    }, 2000);
-}
-
 function saveAttachment(event) {
-    var url = $(this).readAttribute('data-url');
-    if (url) {
-        download(url);
-    }
-    else {
-        $(this).up('ul').select('li[data-url]').each(function (item) {
-            url = $(item).readAttribute('data-url');
-            download(url);
-        });
-    }
+    var div = document.menuTarget;
+    var link = div.select ("a").first ();
+    var url = link.getAttribute("href");
+    var urlAsAttachment = url.replace(/(\/[^\/]*)$/,"/asAttachment$1");
+
+    window.location.href = urlAsAttachment;
 }
 
 /* contacts */
@@ -1988,7 +1984,10 @@ function initMailer(event) {
         initMailboxTree();
         initMessageCheckTimer();
 
-        Event.observe(document, "keydown", onDocumentKeydown);
+        if (Prototype.Browser.Gecko)
+            Event.observe(document, "keypress", onDocumentKeydown); // for FF2
+        else
+            Event.observe(document, "keydown", onDocumentKeydown);
 
         /* Perform an expunge when leaving the webmail */
 //        if (isSafari()) {
@@ -2032,7 +2031,7 @@ function initMailboxTree() {
         node.parentNode.removeChild(node);
     mailboxTree = new dTree("mailboxTree");
     mailboxTree.config.hideRoot = true;
-    mailboxTree.icon.root = ResourcesURL + "/tbtv_account_17x17.png";
+    mailboxTree.icon.root = ResourcesURL + "/tbtv_account_17x17.gif";
     mailboxTree.icon.folder = ResourcesURL + "/tbtv_leaf_corner_17x17.png";
     mailboxTree.icon.folderOpen	= ResourcesURL + "/tbtv_leaf_corner_17x17.png";
     mailboxTree.icon.node = ResourcesURL + "/tbtv_leaf_corner_17x17.png";
@@ -2234,15 +2233,12 @@ function generateMenuForMailbox(mailbox, prefix, callback) {
 function updateMailboxMenus() {
     var mailboxActions = { move: onMailboxMenuMove,
                            copy: onMailboxMenuCopy };
-    var accountsMenus = { move: $$('#moveMailboxMenu li'),
-                          copy: $$('#copyMailboxMenu li') };
 
     for (var key in mailboxActions) {
         for (var i = 0; i < mailAccounts.length; i++) {
             var mailbox = accounts[i];
-            var id = generateMenuForMailbox(mailbox, key + "-" + i,
-                                            mailboxActions[key]);
-            accountsMenus[key][i].submenu = id;
+            generateMenuForMailbox(mailbox, key + "-" + i,
+                                   mailboxActions[key]);
         }
     }
 }
@@ -2492,11 +2488,9 @@ function onMenuLabelNone() {
         });
 }
 
-function onMenuLabelFlag() {
+function _onMenuLabelFlagX(flag) {
     var messages = new Hash();
 
-    var flag = this.readAttribute("data-name");
-    
     if (document.menuTarget.tagName == "DIV")
         // Menu called from message content view
         messages.set(Mailer.currentMessages[Mailer.currentMailbox],
@@ -2508,25 +2502,45 @@ function onMenuLabelFlag() {
             if (row)
                 messages.set(rowID.substr(4),
                              row.getAttribute("labels"));
-        });
+            });
     else
         // Menu called from one selection in messages list view
         messages.set(document.menuTarget.getAttribute("id").substr(4),
                      document.menuTarget.getAttribute("labels"));
-    
+
     var url = ApplicationBaseURL + encodeURI(Mailer.currentMailbox) + "/";
     messages.keys().each(function(id) {
-        var flags = messages.get(id).split(" ");
-        var operation = "add";
-        
-        if (flags.indexOf(flag) > -1)
-            operation = "remove";
-        
-        triggerAjaxRequest(url + id + "/" + operation + "Label?flag=" + flag.asCSSIdentifier(),
-                           messageFlagCallback,
-                           { mailbox: Mailer.currentMailbox, msg: id,
-                             label: operation + flag } );
-    });
+            var flags = messages.get(id).split(" ");
+            var operation = "add";
+
+            if (flags.indexOf("label" + flag) > -1)
+                operation = "remove";
+
+            triggerAjaxRequest(url + id + "/" + operation + "Label" + flag,
+                               messageFlagCallback,
+                               { mailbox: Mailer.currentMailbox, msg: id,
+                                       label: operation + flag } );
+        });
+}
+
+function onMenuLabelFlag1() {
+    _onMenuLabelFlagX(1);
+}
+
+function onMenuLabelFlag2() {
+    _onMenuLabelFlagX(2);
+}
+
+function onMenuLabelFlag3() {
+    _onMenuLabelFlagX(3);
+}
+
+function onMenuLabelFlag4() {
+    _onMenuLabelFlagX(4);
+}
+
+function onMenuLabelFlag5() {
+    _onMenuLabelFlagX(5);
 }
 
 function onMenuToggleMessageFlag(event) {
@@ -2693,33 +2707,14 @@ function onLabelMenuPrepareVisibility() {
 
     var lis = this.childNodesWithTag("ul")[0].childNodesWithTag("li");
     var isFlagged = false;
-
-    // lis is our array of labels, ex:
-    // li
-    // li.seperator
-    // li.label1
-    // li.broccoli
-    // ...
-    for (var i = 2; i < lis.length; i++) {
-
-        // We bind the event handlers if we need to
-        if (lis[i].menuCallback == null) {
-            lis[i].menuCallback = onMenuLabelFlag;
-	    lis[i].on("mousedown", onMenuClickHandler);
-            lis[i].removeClassName("disabled");
-        }
-
-        var flag = lis[i].readAttribute("data-name");
-
-        if (flags[flag]) {
+    for (var i = 1; i < 6; i++) {
+        if (flags["label" + i]) {
             isFlagged = true;
-            lis[i].addClassName("_chosen");
+            lis[1 + i].addClassName("_chosen");
         }
-        else {
-            lis[i].removeClassName("_chosen");
-        }
+        else
+            lis[1 + i].removeClassName("_chosen");
     }
-
     if (isFlagged)
         lis[0].removeClassName("_chosen");
     else
@@ -2824,6 +2819,7 @@ function getMenus() {
                             saveAs, null, null,
                             onMenuDeleteMessage ],
         imageMenu: [ saveImage ],
+        attachmentMenu: [ saveAttachment ],
         messageContentMenu: [ onMenuReplyToSender,
                               onMenuReplyToAll,
                               onMenuForwardMessage,
@@ -2838,19 +2834,21 @@ function getMenus() {
                           onMenuChangeToDraftsFolder,
                           onMenuChangeToTrashFolder  ],
 
-        "label-menu": [ onMenuLabelNone, "-" ],
-
+        "label-menu": [ onMenuLabelNone, "-", onMenuLabelFlag1,
+                        onMenuLabelFlag2, onMenuLabelFlag3,
+                        onMenuLabelFlag4, onMenuLabelFlag5 ],
         "mark-menu": [ onMenuToggleMessageRead, null, null, null, "-", onMenuToggleMessageFlag ],
+// , "-",
+//                        null, null, null ],
 
         searchMenu: [ setSearchCriteria, setSearchCriteria,
                       setSearchCriteria, setSearchCriteria,
                       setSearchCriteria ]
     };
 
-
-    if (typeof mailAccounts != 'undefined') {
-        menus['moveMailboxMenu'] = mailAccounts.collect(function (account) { return account.asCSSIdentifier() });
-        menus['copyMailboxMenu'] = mailAccounts.collect(function (account) { return account.asCSSIdentifier() });
+    var labelMenu = $("label-menu");
+    if (labelMenu) {
+        labelMenu.prepareVisibility = onLabelMenuPrepareVisibility;
     }
 
     var labelMenu = $("label-menu");
@@ -2861,6 +2859,14 @@ function getMenus() {
     var markMenu = $("mark-menu");
     if (markMenu) {
         markMenu.prepareVisibility = onMarkMenuPrepareVisibility;
+    }
+
+    var listMenus = [ "messageListMenu", "messagesListMenu", "messageContentMenu" ];
+    for (var i = 0; i < listMenus.length; i++) {
+        var menu = $(listMenus[i]);
+        if (menu) {
+            menu.prepareVisibility = onMessageListMenuPrepareVisibility;
+        }
     }
 
     var accountIconMenu = $("accountIconMenu");
@@ -2996,14 +3002,20 @@ function stopDragging(event, ui) {
 
 function dropAction(event, ui) {
     var destination = $(this).up("div.dTreeNode");
-    var f;
-    if (ui.helper.hasClass("copy")) {
-        // Message(s) copied
-        f = onMailboxMenuCopy.bind(destination);
+
+    var sourceAct = Mailer.currentMailbox.split("/")[1];
+    var destAct = destination.getAttribute("dataname").split("/")[1];
+    if (sourceAct == destAct) {
+        var f;
+        if (ui.helper.hasClass("copy")) {
+            // Message(s) copied
+            f = onMailboxMenuCopy.bind(destination);
+        }
+        else {
+            // Message(s) moved
+            f = onMailboxMenuMove.bind(destination);
+        }
+
+        f();
     }
-    else {
-        // Message(s) moved
-        f = onMailboxMenuMove.bind(destination);
-    }
-    f();
 }
